2020 - Semestre 2/Análise e Desenvolvimento de Sistemas/3°A/(Equipe 2) VisGeo-ETL/backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
geopandas==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary
py-postgresql==1.2.1
pyshp==2.1.2
Werkzeug==1.0.1
SQLAlchemy==1.3.20
GeoAlchemy2==0.8.4
